complete content moderation for posts and comments

+++ lib/features/posts/db/posts_db.dart
@@ -77,15 +77,35 @@ class PostsDb {
   }) async {
     try {
       final userId = user.userId;
-      await _postsTable.insert({
-        KeyNames.id: postId,
-        KeyNames.content: post,
-        KeyNames.userId: userId,
-        KeyNames.images: images ?? [],
-        KeyNames.can_reposted: canRepost,
-        KeyNames.comments_open: canComment,
-        KeyNames.parent_post: parentId,
-      });
+      final headers = await generateAuthHeaders();
+      // await _postsTable.insert({
+      //   KeyNames.id: postId,
+      //   KeyNames.content: post,
+      //   KeyNames.userId: userId,
+      //   KeyNames.images: images ?? [],
+      //   KeyNames.can_reposted: canRepost,
+      //   KeyNames.comments_open: canComment,
+      //   KeyNames.parent_post: parentId,
+      // });
+
+      final res = await _dio.post(
+        '${appAPI}post',
+        options: Options(headers: headers),
+        data: jsonEncode({
+          KeyNames.post_id: postId,
+          'post': post,
+          KeyNames.userId: userId,
+          KeyNames.images: images ?? [],
+          'can_reposted': canRepost,
+          'can_comment': canComment,
+          'parent_id': parentId,
+          'token': _client.auth.currentSession!.accessToken,
+        }),
+      );
+
+      if (res.statusCode != 200) {
+        throw Exception(res.data);
+      }
 
       final hashtags = extractHashtagKeyword(post).toSet().toList();
       List<SlashEntity> mentions = [];
@@ -206,19 +226,36 @@ class PostsDb {
       List<String> removedMentionsIds = removedMentions.map((m) => m.id).toList();
       final addedMentions =
           mentions.where((mention) => !ogMentions.any((m) => m.id == mention.id)).toList();
-      await _postsTable
-          .update({
-            KeyNames.content: post.content,
-            KeyNames.can_reposted: post.canReposted,
-            KeyNames.comments_open: post.comments_open,
-            KeyNames.updated_at: DateTime.now().toIso8601String(),
-          })
-          .eq(KeyNames.id, post.postId);
+
+      final headers = await generateAuthHeaders();
+
+      final res = await _dio.post(
+        '${appAPI}post-update',
+        options: Options(headers: headers),
+        data: jsonEncode({
+          KeyNames.content: post.content,
+          KeyNames.can_reposted: post.canReposted,
+          KeyNames.comments_open: post.comments_open,
+          KeyNames.id: post.postId,
+          KeyNames.userId: post.userId,
+          'token': _client.auth.currentSession!.accessToken,
+        }),
+      );
+      if (res.statusCode != 200) throw Exception(res.data);
+      // await _postsTable
+      //     .update({
+      //       KeyNames.content: post.content,
+      //       KeyNames.can_reposted: post.canReposted,
+      //       KeyNames.comments_open: post.comments_open,
+      //       KeyNames.updated_at: DateTime.now().toIso8601String(),
+      //     })
+      //     .eq(KeyNames.id, post.postId);
       await Future.wait([
         hashtagDb.removeHashtagsFromPost(removedHashtags, post.postId),
         removeMentionFromPost(removedMentionsIds, post.postId),
         hashtagDb.insertNewHashTag(addedHashtags),
         insertMentions(addedMentions, post.postId),
+        insertEmbedding(id: post.postId, content: post.content, userId: post.userId),
       ]);
       await hashtagDb.insertPostHashTag(addedHashtags, post.postId);
     } catch (e) {
